Put generators into a separate file.
