+++ src/justbytes/_util/misc.py
@@ -23,37 +23,6 @@ import itertools
 import justbases
 
 
-def take_until_satisfied(pred, seq):
-    """
-    Like next(), but yields all values until the first matching value.
-
-    :param bool pred: a predicate, return False if the value is not satisfactory
-    :param seq: a sequence of values
-    """
-    for x in seq:
-        yield x
-        if pred(x):
-            break
-
-def next_or_last(pred, seq, default=None):
-    """
-    Return the first element that matches the predicate or the last element in
-    the seq.
-
-    If seq is empty, return ``default``.
-
-    :param bool pred: a predicate, return False if the value is not satisfactory
-    :param seq: a sequence of values
-    :param default: returned if seq is empty, default is None
-    """
-    for x in seq:
-        if pred(x):
-            return x
-    try:
-        return x # pylint: disable=undefined-loop-variable
-    except NameError:
-        return default
-
 def as_single_number(value, config):
     """
     Returns a rational value as a single number according to the
